mp_server/msg_processor/jenkins_msg/requirements.txt
multi-key-dict==2.0.3
python-jenkins==0.4.14
six==1.10.0